MB-27661 [6/n]: Fix MSVC warnings - C4477, C4805, C4305
This patch addresses the following generated warnings:

C4477 - ‘<function>’ : format string ‘<format-string>’ requires an
        argument of type ‘<type>’, but variadic argument <position>
        has type ‘<type>’
C4805 - operation' : unsafe mix of type 'type' and type 'type' in
	operation. This warning is generated for comparison operations
	between bool and int.
C4305 - identifier' : truncation from 'type1' to 'type2'. The identifier
	is converted to a smaller type, resulting in loss of information.

Reviewed-on: http://review.couchbase.org/89954
Tested-by: Build Bot <build@couchbase.com>

+++ cbsasl/scram-sha/scram-sha.cc
@@ -22,13 +22,14 @@
 #include "cbsasl/util.h"
 
 #include <cbsasl/mechanismfactory.h>
+#include <platform/base64.h>
+#include <platform/random.h>
 #include <cstring>
+#include <gsl/gsl>
 #include <iomanip>
 #include <iostream>
 #include <map>
 #include <memory>
-#include <platform/base64.h>
-#include <platform/random.h>
 #include <set>
 #include <sstream>
 #include <string>
@@ -506,9 +507,11 @@ cbsasl_error_t ScramShaServerBackend::step(const char* input,
     std::string clientproof = iter->second;
     std::string my_clientproof = Couchbase::Base64::encode(getClientProof());
 
-    int fail = cbsasl_secure_compare(clientproof.c_str(), clientproof.length(),
+    int fail = cbsasl_secure_compare(clientproof.c_str(),
+                                     clientproof.length(),
                                      my_clientproof.c_str(),
-                                     my_clientproof.length()) ^user.isDummy();
+                                     my_clientproof.length()) ^
+               gsl::narrow_cast<int>(user.isDummy());
 
     if (fail != 0) {
         if (user.isDummy()) {
